Cleanup: move java implementation to its own directory
Add README file
